requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.8

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client
httpx==0.25.2

# Service discovery and configuration
python-consul2==0.1.5

# Logging
structlog==23.2.0

# Process management (removed psutil due to Windows compilation issues)

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Environment and configuration
python-dotenv==1.0.0

# CORS and middleware
python-multipart==0.0.6

# Template engine (for edge service)
jinja2==3.1.2

# Static file serving
aiofiles==23.2.1

# Date and time handling
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10

# Validation
email-validator==2.1.0

# Async support
asyncio-mqtt==0.16.1

# Health checks
healthcheck==1.3.5

# Rate limiting
slowapi==0.1.9

# Monitoring
prometheus-client==0.19.0

# Error tracking (optional)
sentry-sdk[fastapi]==1.38.0 